fix(schema): add missing tables/columns to SQLite and fix imports
- Fixed ImportError by updating blueprints (onboarding, grandes_contas) to import from correct domain submodules (crud, status, details)
- Fixed 'no such table' SQLite errors by adding config tables (tags_sistema, status_implantacao, etc.) and checklist_finalizacao tables to schema.py
- Added seed data for local development configuration
- Fixed startup order: init_db now runs BEFORE warm_cache
- Fixed config_service query compatibility (Postgres/SQLite) using parameterized queries
- Verified production schema is already up-to-date

+++ cs-onboarding/backend/project/domain/analytics/cancelamentos.py
@@ -161,7 +161,7 @@ def get_cancelamentos_data(cs_email=None, start_date=None, end_date=None, contex
     def pct(p):
         if not tempos_sorted:
             return None
-        idx = max(0, min(len(tempos_sorted) - 1, int(round(p * (len(tempos_sorted) - 1)))))
+        idx = max(0, min(len(tempos_sorted) - 1, round(p * (len(tempos_sorted) - 1))))
         return tempos_sorted[idx]
 
     dist = {"p50": pct(0.5), "p75": pct(0.75), "p90": pct(0.9)}
